fix(cockpit): address code review — extract shared labels, remove dead ref, fix test
