+++ apps/cockpit/src/components/sidebar/navigation-groups.tsx
@@ -4,22 +4,7 @@ import React, { useState } from 'react';
 import type { CockpitManifestEntry } from '@cacheplane/cockpit-registry';
 import type { NavigationProduct } from '../../lib/route-resolution';
 import { toCockpitPath } from '../../lib/route-resolution';
-
-const PRODUCT_LABELS: Record<string, string> = {
-  'deep-agents': 'Deep Agents',
-  'langgraph': 'LangGraph',
-  'render': 'Render',
-  'chat': 'Chat',
-};
-
-
-function stripProductPrefix(title: string): string {
-  const prefixes = ['Deep Agents ', 'LangGraph ', 'Render ', 'Chat '];
-  for (const p of prefixes) {
-    if (title.startsWith(p)) return title.slice(p.length);
-  }
-  return title;
-}
+import { PRODUCT_LABELS, stripProductPrefix } from '../../lib/navigation-labels';
 
 interface NavigationGroupsProps {
   tree: NavigationProduct[];
